requirements.txt
django==4.2.9
django-phonenumber-field[phonenumberslite]==7.3.0
environs==10.2.0
gunicorn==20.1.0
phonenumbers==8.12.2
Pillow==8.3.2
python-dotenv==0.21.1
python-telegram-bot==20.8
pytz==2023.3.post1